add a dropdown button

+++ public/ressources/views/Items/propertyList.php
@@ -248,33 +248,51 @@ $endItem = min($startItem + $itemsPerPage - 1, $totalItems);
                   </div>
                   <div class="table-cell px-4 py-3 p-2 text-sm" style="width: 16.6667%;">
                     <div class="flex items-center space-x-4 text-sm">
-                      <a
-                        class="flex items-center justify-between px-2 py-2 text-sm font-medium leading-5 text-purple-600 rounded-lg dark:text-gray-400 focus:outline-none focus:shadow-outline-gray"
-                        aria-label="Edit" href="index1.php?action=editItem&id=<?php echo $item->getItemId()?>" >
+                      <!-- <button>
+                        <a class="flex items-center justify-between px-2 py-2 text-sm font-medium leading-5 text-purple-600 rounded-lg dark:text-gray-400 focus:outline-none focus:shadow-outline-gray"
+                          aria-label="Edit"
+                          href="index1.php?action=editItem&id=<?php echo $item->getItemId()?>">
 
-                        <svg
-                          class="w-5 h-5"
-                          aria-hidden="true"
-                          fill="currentColor"
-                          viewBox="0 0 20 20"
-                        >
-                          <path
-                            d="M13.586 3.586a2 2 0 112.828 2.828l-.793.793-2.828-2.828.793-.793zM11.379 5.793L3 14.172V17h2.828l8.38-8.379-2.83-2.828z"
-                          ></path>
+                          <svg class="w-5 h-5" aria-hidden="true" fill="currentColor" viewBox="0 0 20 20">
+                            <path d="M13.586 3.586a2 2 0 112.828 2.828l-.793.793-2.828-2.828.793-.793zM11.379 5.793L3 14.172V17h2.828l8.38-8.379-2.83-2.828z"></path>
+                          </svg>
+                        </a>
+                      </button>
+                      <button>
+                        <a
+                          class="flex items-center justify-between px-2 py-2 text-sm font-medium leading-5 text-purple-600 rounded-lg dark:text-gray-400 focus:outline-none focus:shadow-outline-gray"
+                          aria-label="Delete" onclick="return confirm('voulez vous vraiment supprimer ce utilisateur')" 
+                          href="index1.php?action=destroyItem&id=<?php echo $item->getItemId()?>" >
+                          <svg class="w-5 h-5" aria-hidden="true" fill="currentColor" viewBox="0 0 20 20" >
+                            <path
+                              fill-rule="evenodd"
+                              d="M9 2a1 1 0 00-.894.553L7.382 4H4a1 1 0 000 2v10a2 2 0 002 2h8a2 2 0 002-2V6a1 1 0 100-2h-3.382l-.724-1.447A1 1 0 0011 2H9zM7 8a1 1 0 012 0v6a1 1 0 11-2 0V8zm5-1a1 1 0 00-1 1v6a1 1 0 102 0V8a1 1 0 00-1-1z"
+                              clip-rule="evenodd" >
+                            </path>
+                          </svg>
+                        </a>
+                      </button> -->
+                      <div class="relative ">
+                        <button data-dropdown-trigger="hover" class="relative z-10 align-middle text-purple-600  rounded-md focus:outline-none focus:shadow-outline-purple" type="button" onclick="toggleDropdown(this)" data-item-id="<?php echo $item->getItemId() ?>">
+                        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-more-vertical">
+                          <circle cx="12" cy="12" r="1"></circle>
+                          <circle cx="12" cy="5" r="1"></circle>
+                          <circle cx="12" cy="19" r="1"></circle>
                         </svg>
-                      </a>
-                      <a
-                        class="flex items-center justify-between px-2 py-2 text-sm font-medium leading-5 text-purple-600 rounded-lg dark:text-gray-400 focus:outline-none focus:shadow-outline-gray"
-                        aria-label="Delete" onclick="return confirm('voulez vous vraiment supprimer ce utilisateur')" 
-                        href="index1.php?action=destroyItem&id=<?php echo $item->getItemId()?>" >
-                        <svg class="w-5 h-5" aria-hidden="true" fill="currentColor" viewBox="0 0 20 20" >
-                          <path
-                            fill-rule="evenodd"
-                            d="M9 2a1 1 0 00-.894.553L7.382 4H4a1 1 0 000 2v10a2 2 0 002 2h8a2 2 0 002-2V6a1 1 0 100-2h-3.382l-.724-1.447A1 1 0 0011 2H9zM7 8a1 1 0 012 0v6a1 1 0 11-2 0V8zm5-1a1 1 0 00-1 1v6a1 1 0 102 0V8a1 1 0 00-1-1z"
-                            clip-rule="evenodd" >
-                          </path>
-                        </svg>
-                      </a>
+                        </button>
+                        <!-- Dropdown menu -->
+                        <div id="dropdownDelay-<?php echo $item->getItemId() ?>" class="absolute z-20 right-0 w-56 p-2 mt-2 space-y-2 hidden text-gray-600 bg-white border border-gray-100 rounded-md shadow-md dark:text-gray-300 dark:border-gray-700 dark:bg-gray-700">
+                          <ul class="py-2 text-sm text-gray-700 dark:text-gray-200" aria-labelledby="dropdownDelayButton">
+                            <li><a href="index1.php?action=editItem&id=<?php echo $item->getItemId()?>"
+                             class="block px-4 py-2 hover:bg-gray-100 dark:hover:bg-gray-600 dark:hover:text-white">Modifier</a></li>
+                            <li><a href="#" class="block px-4 py-2 hover:bg-gray-100 dark:hover:bg-gray-600 dark:hover:text-white">Réserver</a></li>
+                            <li><a onclick="return confirm('voulez vous vraiment supprimer ce utilisateur')" href="index1.php?action=destroyItem&id=<?php echo $item->getItemId()?>" 
+                             class="block px-4 py-2 hover:bg-gray-100 dark:hover:bg-gray-600 dark:hover:text-white">Supprimer</a></li>
+                          </ul>
+                        </div>
+                      </div>
+            
+
                     </div>
                   </div>
               </div> 
@@ -336,7 +354,20 @@ $endItem = min($startItem + $itemsPerPage - 1, $totalItems);
   <?php
   include_once 'assets/js/global.js';
   ?>
+
+      
+    
+
+    // document.addEventListener('click', function(event) {
+    //   if (!optionsList.contains(event.target) && event.target !== optionsButton) {
+    //     optionsList.classList.add('hidden');
+    //   }
+    // });
+  // });
+
 </script>
+
+
 <?php
 $content = ob_get_clean();
 include_once 'ressources/views/layout.php';
